Social media icons: Restrict logo CSS classes to the scenarios they're meant to be used in.
Supported scenarios:
* Children of parent elements that use the "followus" class.
* Elements that use the "icon" class.

Prevents conflicts with plugins that use classes with identical names to the logo classes.

Example: The multimedia player's YouTube examples use a "youtube" class that has nothing to do with the "youtube" social media icon class. The latter class was previously causing GCWeb implementations of the player to use a repeating pattern of YouTube logos in its background.