Advanced service initiation pages: Updated ToC list's class in most pages.
The first English page's ToC uses a class called "gc-stp-stp", but its French counterpart and pages 2-6 were using "mrgn-tp-md brdr-bttm". As a result of that inconsistency, the top of the right panel was "sticking" against the ToC's bottom border in all of the latter pages.

This commit updates all affected pages' ToCs to use the "gc-stp-stp" class, which increases consistency and resolves the aforementioned spacing issue.