Theme: Fixed header FIP image's incomplete alt text.
A bilingual FIP image is situated in the top-right of all of GCWeb's pages. Until now, in all standard/server message page templates, the image's alt text was unilingual, which isn't an accurate representation of the text in the image.

This commit adds extra handlebars conditions to make the FIP's alt text bilingual depending on the current page language.