Theme: Fixed incorrect FIP alt text in arm's length institution pages.
The Government of Canada FIP is used as a sample image throughout arm's-length institution page templates. Until now, it was using "Department name"/"Nom du département" as alt text, which is incorrect.

This commit fixes it by removing the "arms-image-alt" i18n variable, replacing all references to it with "tmpl-gc-sig" and adding extra is/else logic to account for the fact that the FIP is bilingual.