Search: Fixed search button layout issues.
In medium/large views, the search field/button's layout used to rely on space/line break/tab characters being present in-between their elements in HTML markup. Their presence "filled" in the left side of the search button to make it look square-shaped. But in pages that use minified HTML, the space disappeared, which hindered the button's layout.

It turns out that the search feature's small view and under SCSS file contained selectors that eliminated the space in noscript/wbdisable mode.

This commit moves the aforementioned selectors into a screen SCSS file, adds left/right margins on the button to retain its square shape and adjusts the button's height to match the search field. Combined, these allow the button to always remain square-shaped and flush with its input field.

Related to wet-boew/wet-boew#8061.