Campaign: Added second example page and related updates.
This is a finalized version of an alternate campaign page template that won a hackathon competition ESDC internally hosted a few months ago. TBS is on board with it and @shawnthompson previously reviewed it. It's since received a few minor CSS/HTML updates to fix some issues, but hasn't fundamentally changed since it was approved/reviewed.

Specific changes:
* Added a second campaign page example page.
* Added related SCSS styles (won't conflict with the original campaign page template).
* Added supporting images.
* Updated the index page. The original campaign page template is now presented as example 1, whereas the second one is presented as example 2.

Thanks @GormFrank for all your help with this template's HTML/CSS :)!