Skip links: Changed "About this site" to "About government".
Prior to #1132, the first heading within the wb-info ID's container was called "About this site". As a part of #1132's changes, the footer was restructured into two sections of links ("About government" and "About this site"). As a result of those changes, wb-info's first heading changed to "About government"... but its skip link was never updated. Ever since then, the "Skip to 'About this site'" skip link has technically been pointing to the "About government" heading.

This commit fixes that inconsistency by renaming the "Skip to 'About this site'" link to "Skip to 'About government'".