Report a problem: Fixed dropdown box's width.
The report a problem dropdown box's width was previously thinner than the summary button that directly precedes it. Bootstrap's "row" class was being misused to increase its width, but it still wasn't enough. The "row" class' negative left/right margins differ from the summary's margins.

This commit makes the dropdown box's width consistent with the preceding summary. It removes the unnecessary "row" class and overrides the "well" class' margins to rely on the same SCSS variable that's used to generate the summary element's left/right margins. These changes are backwards-compatible with GCWeb 4.0.26 and under.