Promotion1 include: Removed "row" div.
The promotion1 include only contains a carousel. It doesn't use a grid layout, so the "row" class was being misused. That class was also causing the include's carousel to become wider than the content area and "touch" that area's far left/right edges. It was especially noticeable in extra-small view and under.

This commit removes the "row" div and affects the following page templates:
* Campaign (example 1)
* Event
* Home
* News (English-only)