Theme: Removed all references to the "gc-drmt" class.
That class wasn't associated to any SCSS styles. It also wasn't used consistently across either GCWeb nor real-world implementations of that theme (including the MWS).

If it solely existed as a placeholder class to make it possible to easily modify the layout of all doormat links in the future, it was unable to fully achieve that objective since it was missing from many pages that use those kinds of links.