Remove GC priorities and improved arms length footer
Removed template variable and assets related to the GC priority feature. Deprecated CSS that support the old markup of the arms length and the removed GC priority.

The arms length footer was moved in it's own file instead being integrated with the gc priorities (feature.hbs) include. It's markup was improved and it's old markup are now deprecated.