Theme: Hid expanded report a problem dropdowns when printing.
Previously, if a user expanded the "Report a problem on this page" dropdown in a browser and subsequently printed the page they were on, the dropdown's expanded content would appear in the printout. However, the expanded content served no purpose in that context. Furthermore, since the "Report a problem on this page" button is hidden when printing, the purpose of the expanded content wasn't clear.

This change prevents the expanded content from ever appearing in printouts.