Video example: Use repo's copies of poster images
This is another take on what was done in #1576.

GCWeb contains copies of most (all?) of the WET repo's plugin demo pages, documentation and supporting files. So it makes more sense for the video example page's poster image references to be relative to GCWeb instead of referencing an external repo (wet-boew).