Splash page: Remove French meta description var
Fixes the following error in the htmllint:all task (and official HTML validator):
"A document must not include more than one “meta” element with its “name” attribute set to the value “description”."

The WHATWG's HTML spec (which is now the W3C's canonical spec) currently disallows multiple meta descriptions on the same page:
https://html.spec.whatwg.org/multipage/semantics.html#meta-description

The W3C's archived HTML spec used to allow it (via w3c/html#1161). A similar change has yet to be proposed for the WHATWG's spec.

Related to #1598 and #1599.