Theme: Reset the .label class' line height
The form font resizing exclusions that were added in #1449 didn't cover the .label class. This adds it in.

As a result, this prevents the form validation plugin's field error messages from overflowing on top of labels and legends under certain circumstances. If a class was set on a label or legend to manipulate its font (such as .h1-.h6), it was causing a significant line-height mismatch between the label/legend and its associated form validation error message. That behaviour was inadvertently introduced in #1449.