Move GC brand text to img alt
Because the GC span text is removed, the RDFa property  needs to be supplemented.

The GC logo is only in one language because logos are an exception from verbose alternative text. The meaning is sufficiently represented and does not need alternative text beyond that of the chosen language of the page.
https://github.com/wet-boew/GCWeb/issues/1670