Field flow: Use "chkbxrdio-grp" class for fieldsets
This is how fieldsets with normal-sized legends are supposed to be coded in WET.

Field flow was previously misusing the "h5" class in its fieldset legends to get a similar visual effect. But that method has error message overlapping issues in modern versions of GCWeb and won't benefit from any solutions for #1634 (like #1654 and #1676).