GC tables: Support adjacent tbodies
Bootstrap's "tables" class adds a 2px border between adjacent tbody elements.

But it looked out of place in GC tables. That component adds a bottom margin to every row... which caused the adjacent tbody's top border to look like it was "sticking" to its first row.

This resolves it by adding a top margin to the adjacent tbody's first row to prevent the sticking. That allows the top border to look like a real separator. Also adds a second tbody to the first GC tables example.

Spinoff of #1843.