Subway map menu: Use H2 for document navigation
The document navigation section was previously using an H3 heading... which didn't make structural sense (it's preceded by an H1 heading and doesn't function as a subsection).

The H3 was probably a leftover from the advanced service initiation page template. That template intentionally used an H3 for its document navigation section.