Build: Add final newline to generated JSON files
The concat:components/templates/sites tasks were previously generating JSON files that lacked a final newline character in the _data folder. That contradicted the repo's .editorconfig file.

This resolves it by adding "\n" to the end of the concat tasks' footer variables. Grunt also seems to treat it in a platform-agnostic way (generates CRLF newlines in Windows).