Footer: Update top of page link's destination
GCWeb's "Top of page"/"Haut de la page" link was previously pointing to #wb-cont. That's the primary H1 heading's ID (usually top of the content area). The link text was misleading since it didn't actually go to the top of the page.

This resolves it by changing the link's destination to #wb-tphp. That ID is fit for purpose. Older versions of WET also used top of page links that pointed to that ID.