Page details: Fix links to report a problem's AJAX fragment #1989
On GitHub Pages, most of the repo's pages are currently unable to load the report a problem widget's AJAX fragment due to the feedbackPath variable beginning with "/sites/...". It probably works fine in local instances of Jekyll, but fails on GH Pages. Why? Because GCWeb's GH Pages site uses "/GCWeb/" as its top-level folder.

This fixes it by adding Liquid's relative_url filter in the Data Ajax plugin's default reference to feedbackPath. Using it causes site.baseurl to be prepended to feedbackPath. That should produce "/sites/..." in local Jekyll instances and "/GCWeb/sites/..." in GH Pages.

The relative_url filter isn't applied to pages that override the feedbackPath variable. Therefore, overrides must use paths that are relative to the current page.

Related changes:
* Page details documentation:
  * English version:
    * Fix custom path for feedbackPath that wasn't relative to the page
    * Add missing data-ajax-replace attribute in the "Page details without 'Share this page' button" code sample
  * French version:
    * Fix greater than signs (>) that weren't coded as named entities

Supersedes part of #1975.