Field flow: Detect anchors via firstElementChild
Anchors were previously detected by evaluating firstChild's node name... which only worked if the very first node was an anchor element.

As a result, if the first node was something other than an element (e.g. comment or text node with empty space)... they'd be evaluated instead of the real anchor and consequently mess up the redirection action.

This resolves it by evaluating firstElementChild instead of firstChild. Also adds a comment and empty space to some of the redirection examples.