Update and-or.css
Added min-height to all side by side items to 3em; to exceed the height of the and/or CSS text and container (::before), as well as adjust for items where there is padding for last child element (e.g. alerts).   This will not effect any current visual look other than may provide a better visual separation (border line) for items with a small amount of content.