Clipboard plugin (méli-mélo): Restore space indents
#2411 accidentally-introduced tab indents into the plugin's JS file.

It was caused by a clash of EditorConfig settings (which apply GCWeb's basic linting settings to all méli-mélo files) vs ESLint (which excludes méli-mélo). The added indents adhered to the EditorConfig settings, but the file itself was using a different kind of indent style (groups of 4 spaces).

This changes the tab characters back into spaces.