Update steps.css
Clear and simple explanation of the change/update:

Additional CSS and example added to support basic use (for numbers 2 to 10) of the start attribute for ordered (ol) lists as the CSS design uses a CSS count method to display the numbers and not the generated numbers from the html element. 

Future design could include a javascript to allow any start number to exist.  

The impact on the sponsored department (CRA) for that change/update:

No impact

The impact on the public for that change/update:

More flexibility of using the design in various situations.