Patch - Steps Méli-mélo: Update steps.css (#2441)
* Update steps.css

Clear and simple explanation of the change/update:

Additional CSS and example added to support basic use (for numbers 2 to 10) of the start attribute for ordered (ol) lists as the CSS design uses a CSS count method to display the numbers and not the generated numbers from the html element. 

Future design could include a javascript to allow any start number to exist.  

The impact on the sponsored department (CRA) for that change/update:

No impact

The impact on the public for that change/update:

More flexibility of using the design in various situations.

* Update index.html

Added extended functionality example for use of start attribute for steps

* Update index-fr.html

Added extended functionality example using start attribute for steps

* Update steps-doc-en.html

Added start attribute code

* Update steps-doc-fr.html

Added code snippet for start attribute

* Update meta.md

Added updated feature to list

* Méli-mélo 2021-05-steps - Add working example for custom starting step

* Méli-mélo 2021-05-steps - Typo fix for custom starting steps

* Méli-mélo 2021-05-steps - Typo fix for custom starting steps

---------

Co-authored-by: Pierre Dubois <duboisp2@gmail.com>