Theme: Fix placeholder image references
Some of GCWeb's code samples from recent years were referencing placeholder images from placeholder.com and placehold.it. But those sites underwent changes in ownership in March 2023 and consequently ceased serving placeholder images.

This deals with it by replacing all references to via.placeholder.com and placehold.it with dummyimage.com (and swaps ? characters with & where custom text is specified to fit the latter's URL writing scheme).

GCWeb has historically used dummyimage.com and that site still exists to this day.

Left the following references to placeholder.com "as-is" in the following WCAG assessments since they're meant to represent specific points in time:
* templates/institutional-landing/reports/a11y-1.json
* templates/institutional-landing/reports/a11y-2.json